+++ spec/unit/xml/mapping_extensions/namespace_spec.rb
@@ -1,0 +1,13 @@
+require 'spec_helper'
+
+module XML
+  module MappingExtensions
+    describe Namespace do
+      describe '#set_default_namespace' do
+        it 'sets the default namespace'
+
+        # TODO: Other tests
+      end
+    end
+  end
+end